ai_server/pyproject.toml
[tool.poetry]
name = "onsikgu-ai-server"
version = "2.0.0"
description = "온식구 AI 서버 - 가족 질문 생성 및 답변 분석"
authors = ["Minsu Park <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
openai = "^1.57.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
pydantic-settings = "^2.8.0"
httpx = "^0.28.1"
chromadb = "^0.4.22"
numpy = "<2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langsmith = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"
pytest-dotenv = "^0.5.2"
ruff = "^0.8.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # imported but unused
"tests/**/*.py" = ["B018"]  # useless expression

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
env_files = [".env"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
